fix config CI and review feedback
