+++ crates/logfwd-config/src/lib.rs
@@ -37,10 +37,10 @@ pub use types::{
     HttpTypeConfig, InputConfig, InputType, InputTypeConfig, JournaldBackendConfig,
     JournaldInputConfig, JournaldTypeConfig, JsonlEnrichmentConfig, K8sPathConfig,
     LokiOutputConfig, NullOutputConfig, OtlpOutputConfig, OtlpProtobufDecodeModeConfig,
-    OtlpProtocol, OtlpTypeConfig, OutputConfigV2, OutputType, PipelineConfig, S3InputConfig,
-    S3TypeConfig, SensorTypeConfig, ServerConfig, SourceMetadataStyle, StaticEnrichmentConfig,
-    StdoutOutputConfig, StorageConfig, TcpOutputConfig, TcpTypeConfig, UdpOutputConfig,
-    UdpTypeConfig,
+    OtlpProtocol, OtlpTypeConfig, OutputConfigV2, OutputType, PipelineConfig, S3CompressionConfig,
+    S3InputConfig, S3TypeConfig, SensorTypeConfig, ServerConfig, SourceMetadataStyle,
+    StaticEnrichmentConfig, StdoutOutputConfig, StorageConfig, TcpOutputConfig, TcpTypeConfig,
+    UdpOutputConfig, UdpTypeConfig,
 };
 pub use validate::validate_host_port;
