+++ dev-docs/research/source-metadata-attachment-plan.md
@@ -37,7 +37,7 @@ PR #2317 landed the sidecar spine:
 - `source_metadata: fastforward` materializes `__source_id` as nullable
   `UInt64`.
 - Public styles materialize source paths as normal columns: `file.path` for
-  ECS/Beats, `log.file.path` for OTel, and `file` for Vector.
+  ECS, `log.file.path` for OTel, and `file` for Vector.
 - Public source path styles are accepted only for inputs that expose path
   snapshots. File tailing exposes filesystem paths. S3 exposes object keys when
   a public style is configured. This guard avoids replacing a payload
